requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
email-validator==2.2.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.14.0
python-multipart==0.0.12
pandas==2.2.3
openpyxl==3.1.5
pdfplumber==0.11.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==43.0.3
httpx==0.27.2
python-dotenv==1.0.1
